requirements.txt
# Base
Django==1.10.5
psycopg2==2.6.2
django-crispy-forms

# Dev
django_debug_toolbar